move cas flow to an own file and split to smaller functions
